deconstify VG_NodeOps
git-svn-id: https://dev.csoft.net/agar/trunk@7959 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ vg/vg_arc_tool.c
@@ -68,7 +68,7 @@ MouseButtonDown(void *p, VG_Vector vPos, int button)
 				pCenter = VG_PointNew(vg->root, vPos);
 			}
 			t->vaCur = VG_ArcNew(vg->root, pCenter, 1.0f,
-			    0.0f, VG_PI);
+			    0.0f, (float)VG_PI);
 			AdjustRadius(t->vaCur, vPos);
 		} else {
 			AdjustRadius(t->vaCur, vPos);
